fix(samples): Teaching-code cleanups from nightly review (#1240)
Address teaching-code anti-patterns surfaced by nightly review #1208.
Samples are copied by users, so these style/convention issues matter.

- Simulation: EnemyShootingSystem now queries a semantic [TagComponent]
  RangedAttacker instead of gating on the rendered symbol ('@'). Tough
  enemies are tagged at spawn and given a Cooldown so the "enemies shoot
  back" behavior actually runs (teaches Separation of Concerns).
- AIProximity / CollisionDetection / Aot: split components and systems
  into Components.cs / Systems.cs under a file-scoped namespace, matching
  the Graphics/Input sample pattern, so public types no longer land in
  the global namespace.
- AIProximity: removed the unused guardEntity parameter and its
  "reserved for future" comment from CheckForThreats.
- MassSimulation: removed the computed-then-discarded per-frame timing.
- RenderingCulling: removed the unused Camera.IsActive field (Camera is
  now a tag component, spawned via WithTag<Camera>()).
- UI/MathNodes: replaced NumberNode.Initialize's misleading
  "Could add component" stub comment with an accurate note.
- StringTags: removed the AIBehavior component that was spawned but never
  read (the sample drives state entirely off string tags).

Closes #1240
